Added control on Missing parameters

+++ EshopELIS/src/main/java/com/elismarket/eshop/eshopelis/service/UtenteServiceImpl.java
@@ -1,8 +1,7 @@
 package com.elismarket.eshop.eshopelis.service;
 
 import com.elismarket.eshop.eshopelis.dto.*;
-import com.elismarket.eshop.eshopelis.exception.ExceptionPhrases;
-import com.elismarket.eshop.eshopelis.exception.UtenteException;
+import com.elismarket.eshop.eshopelis.exception.*;
 import com.elismarket.eshop.eshopelis.helper.FeedbackHelper;
 import com.elismarket.eshop.eshopelis.helper.PagamentoHelper;
 import com.elismarket.eshop.eshopelis.helper.ProdottoHelper;
@@ -11,6 +10,7 @@ import com.elismarket.eshop.eshopelis.model.*;
 import com.elismarket.eshop.eshopelis.repository.UtenteCrud;
 import com.elismarket.eshop.eshopelis.service.interfaces.UtenteService;
 import com.elismarket.eshop.eshopelis.utility.Checkers;
+import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -85,9 +85,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param utenteDTO {@link UtenteDTO UtenteDTO} with updated fields
      * @return updated Utente
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public UtenteDTO updateUtente(Long userId, UtenteDTO utenteDTO) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
 
         duplicateChecker(utenteDTO);
 
@@ -151,9 +155,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param id of the {@link Utente Utente}
      * @return HTTP 200 if deleted successfully, else 500
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Boolean removeUtente(Long id) {
+        if (Objects.isNull(id))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
         if (!utenteCrud.existsById(id))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         utenteCrud.deleteById(id);
@@ -208,9 +216,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param mail of the {@link Utente Utente}
      * @return retrieved Utente
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public UtenteDTO getByMail(String mail) {
+        if (Objects.isNull(mail) || Strings.isBlank(mail) || Strings.isEmpty(mail))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
         if (Objects.isNull(utenteCrud.findByMail(mail)))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return Utente.to(utenteCrud.findByMail(mail));
@@ -222,9 +234,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param username of the {@link Utente Utente}
      * @return retrieved Utente
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public UtenteDTO getByUser(String username) {
+        if (Objects.isNull(username) || Strings.isBlank(username) || Strings.isEmpty(username))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
         if (Objects.isNull(utenteCrud.findByUsername(username)))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return Utente.to(utenteCrud.findByUsername(username));
@@ -236,9 +252,17 @@ public class UtenteServiceImpl implements UtenteService {
      * @param siglaResidenza of the {@link Utente Utente}
      * @return retrieved Utente
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
+     * @throws UtenteException with {@link ExceptionPhrases#INCONSISTENT_SIGLA INCONSISTENT_SIGLA} message
      */
     @Override
     public UtenteDTO getBySigla(Integer siglaResidenza) {
+        if (Objects.isNull(siglaResidenza))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
+        if (Checkers.siglaChecker(siglaResidenza))
+            throw new UtenteException(INCONSISTENT_SIGLA.name());
+
         if (Objects.isNull(utenteCrud.findBySiglaResidenza(siglaResidenza)))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return Utente.to(utenteCrud.findBySiglaResidenza(siglaResidenza));
@@ -250,9 +274,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param userId of the {@link Utente Utente}
      * @return retrieved Utente
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public UtenteDTO getById(Long userId) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
         if (!utenteCrud.existsById(userId))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return Utente.to(utenteCrud.findById(userId).orElseThrow(() -> new UtenteException(CANNOT_FIND_ELEMENT.name())));
@@ -268,7 +296,9 @@ public class UtenteServiceImpl implements UtenteService {
      */
     @Override
     public UtenteDTO getLoginUtente(String username, String password) {
-        if (Objects.isNull(username) || Objects.isNull(password))
+        if (Objects.isNull(username) || Objects.isNull(password) ||
+                Strings.isEmpty(username) || Strings.isBlank(username) ||
+                Strings.isEmpty(password) || Strings.isBlank(password))
             throw new UtenteException(MISSING_PARAMETERS.name());
 
         Utente u = utenteCrud.findByUsernameAndPassword(username, Utente.hashPassword(password));
@@ -283,9 +313,13 @@ public class UtenteServiceImpl implements UtenteService {
      * @param userId of the {@link Utente Utente} to logout
      * @return HTTP status 200 if all went good, otherwise 500
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Boolean getLogout(Long userId) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+
         Utente u = utenteCrud.findById(userId).orElseThrow(() -> new UtenteException(CANNOT_FIND_ELEMENT.name()));
         u.setLogged(Boolean.FALSE);
         utenteCrud.saveAndFlush(u);
@@ -298,10 +332,17 @@ public class UtenteServiceImpl implements UtenteService {
      * @param userId      of the {@link Utente Utente}
      * @param feedbackDTO {@link FeedbackDTO FeedbackDTO} to link to the Utente
      * @return Added Feedback
-     * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException   with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException   with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
+     * @throws FeedbackException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Feedback addFeedbackToUser(Long userId, FeedbackDTO feedbackDTO) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+        if (Objects.isNull(feedbackDTO))
+            throw new FeedbackException(MISSING_PARAMETERS.name());
+
         if (!utenteCrud.existsById(userId))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return feedbackHelper.addFeedbackToUser(userId, feedbackDTO);
@@ -314,9 +355,16 @@ public class UtenteServiceImpl implements UtenteService {
      * @param propostaDTO {@link PropostaDTO PropostaDTO} to link to the Utente
      * @return Added Proposta
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
+     * @throws PropostaException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Proposta addPropostaToUser(Long userId, PropostaDTO propostaDTO) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+        if (Objects.isNull(propostaDTO))
+            throw new PropostaException(MISSING_PARAMETERS.name());
+
         if (!utenteCrud.existsById(userId))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return propostaHelper.addPropostaToUser(userId, propostaDTO);
@@ -329,9 +377,16 @@ public class UtenteServiceImpl implements UtenteService {
      * @param prodottoDTO {@link ProdottoDTO ProdottoDTO} to link to the Utente
      * @return Added Prodotto
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
+     * @throws ProdottoException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Prodotto addProdottoToUser(Long userId, ProdottoDTO prodottoDTO) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+        if (Objects.isNull(prodottoDTO))
+            throw new ProdottoException(MISSING_PARAMETERS.name());
+
         if (!utenteCrud.existsById(userId))
             throw new UtenteException(CANNOT_FIND_ELEMENT.name());
         return prodottoHelper.addProdottoToUser(userId, prodottoDTO);
@@ -344,9 +399,16 @@ public class UtenteServiceImpl implements UtenteService {
      * @param pagamentoDTO {@link PagamentoDTO PagamentoDTO} to link to the Utente
      * @return Added Pagamento
      * @throws UtenteException with {@link ExceptionPhrases#CANNOT_FIND_ELEMENT CANNOT_FIND_ELEMENT} message
+     * @throws UtenteException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
+     * @throws PagamentoException with {@link ExceptionPhrases#MISSING_PARAMETERS MISSING_PARAMETERS} message
      */
     @Override
     public Pagamento addPagamentoToUser(Long userId, PagamentoDTO pagamentoDTO) {
+        if (Objects.isNull(userId))
+            throw new UtenteException(MISSING_PARAMETERS.name());
+        if (Objects.isNull(pagamentoDTO))
+            throw new PagamentoException(MISSING_PARAMETERS.name());
+
         if (utenteCrud.existsById(userId))
             return pagamentoHelper.addPagamentoToUser(userId, pagamentoDTO);
